+++ settings/bash.bashrc.d/06_less-colors.sh
@@ -1,0 +1,11 @@
+# shellcheck shell=bash
+# $'...' ANSI-C quoting below is a bashism - not POSIX-safe, hence bash.bashrc.d only.
+
+# Colour for manpages in less, makes manpages a little easier to read.
+export LESS_TERMCAP_mb=$'\E[01;31m'
+export LESS_TERMCAP_md=$'\E[01;31m'
+export LESS_TERMCAP_me=$'\E[0m'
+export LESS_TERMCAP_se=$'\E[0m'
+export LESS_TERMCAP_so=$'\E[01;44;33m'
+export LESS_TERMCAP_ue=$'\E[0m'
+export LESS_TERMCAP_us=$'\E[01;32m'