+++ settings/bash.bashrc.d/80_docker.sh
@@ -1,0 +1,4 @@
+# shellcheck shell=sh
+if command -v docker > /dev/null 2>&1; then
+    alias docker="sudo docker"
+fi