+++ settings/bash.bashrc.d/90_shutdown.sh
@@ -1,0 +1,3 @@
+# shellcheck shell=sh
+alias shutdown="sudo shutdown"
+alias reboot="sudo reboot"